telewall/telewall/requirements.txt
# compatible with py3
future==0.15.2
six==1.10.0

# core
statu==0.3.1
phonenumbers==7.1.0
timeout-decorator==0.3.2
py-lru-cache==0.1.4

# asterisk
ari==0.1.3

# web
Flask==0.10.1
requests==2.8.1

# sql
SQLAlchemy==1.0.9

# testing
nose==1.3.7
nosexcover==1.0.10
pylint==1.4.4
mock==1.3.0

# ui
RPLCD==0.3.0
RPi.GPIO==0.5.11